requirements.txt
djangorestframework
djangorestframework-simplejwt
pillow
qrcode
django-filter
python-dotenv
gunicorn
requests
   # or paystack integration
